+++ packages/core/src/config/index.ts
@@ -135,7 +135,10 @@ export async function normalizeUserCompilationConfig(
   }
 
   // we should not deep merge compilation.input
-  if (userConfig.compilation?.input) {
+  if (
+    userConfig.compilation?.input &&
+    Object.keys(userConfig.compilation.input).length > 0
+  ) {
     // Add ./ if userConfig.input is relative path without ./
     const input: Record<string, string> = {};
 
@@ -286,6 +289,7 @@ export async function resolveInlineConfig(
       const farmConfig = mergeUserConfig(config, options);
       if (config) {
         userConfig = parseUserConfig(farmConfig);
+
         // if we found a config file, stop searching
         break;
       }