+++ packages/core/src/index.ts
@@ -21,6 +21,7 @@ import { DevServer } from './server/index.js';
 import { FileWatcher } from './watcher/index.js';
 import type { FarmCLIOptions } from './config/types.js';
 import { Config } from '../binding/index.js';
+import { compilerHandler } from './utils/build.js';
 
 export async function start(
   inlineConfig: FarmCLIOptions & UserConfig
@@ -46,8 +47,8 @@ export async function start(
       process.exit(1);
     }
 
-    const fileWatcher = new FileWatcher(config.root, devServer.config.hmr);
-    fileWatcher.watch(devServer, {});
+    const fileWatcher = new FileWatcher(devServer, normalizedConfig);
+    fileWatcher.watch();
   }
 }
 
@@ -61,22 +62,7 @@ export async function build(
     'production'
   );
 
-  const start = Date.now();
-  const compiler = new Compiler(normalizedConfig);
-  compiler.removeOutputPathDir();
-  if (userConfig.compilation?.watch) {
-    createFileWatcher(userConfig.root, compiler, normalizedConfig);
-  } else {
-    await compiler.compile();
-    compiler.writeResourcesToDisk();
-    logger.info(
-      `⚡️ Build completed in ${chalk.green(
-        `${Date.now() - start}ms`
-      )}! Resources emitted to ${chalk.green(
-        normalizedConfig.config.output.path
-      )}.`
-    );
-  }
+  createBundleHandler(normalizedConfig);
 }
 
 export async function preview(options: FarmCLIOptions): Promise<void> {
@@ -140,19 +126,23 @@ export async function watch(
     userConfig,
     'production'
   );
-  const compiler = new Compiler(normalizedConfig);
-  createFileWatcher(userConfig.root, compiler, normalizedConfig);
+
+  createBundleHandler(normalizedConfig, true);
 }
 
-export function createFileWatcher(
-  watcherDirPath: string,
-  compiler: Compiler,
-  normalizedConfig: Config
+export async function createBundleHandler(
+  normalizedConfig: Config,
+  watchMode = false
 ) {
-  const outDir = normalizedConfig.config.output.path;
-  const outDirRegex = new RegExp(`^.*${outDir}.*$`);
-  const fileWatcher = new FileWatcher(watcherDirPath, {
-    ignores: ['**/{.git,node_modules}/**', outDirRegex]
-  });
-  fileWatcher.watch(compiler, normalizedConfig);
+  const compiler = new Compiler(normalizedConfig);
+  await compilerHandler(async () => {
+    compiler.removeOutputPathDir();
+    await compiler.compile();
+    compiler.writeResourcesToDisk();
+  }, normalizedConfig);
+
+  if (normalizedConfig.config?.watch || watchMode) {
+    const watcher = new FileWatcher(compiler, normalizedConfig);
+    watcher.watch();
+  }
 }