+++ packages/core/src/utils/common.ts
@@ -25,3 +25,16 @@ export function tryStatSync(file: string): fs.Stats | undefined {
     // Ignore errors
   }
 }
+
+// Taken from https://github.com/sindresorhus/slash/blob/main/index.js (MIT)
+export function slash(path: string) {
+  const isExtendedLengthPath = /^\\\\\?\\/.test(path);
+  // eslint-disable-next-line no-control-regex
+  const hasNonAscii = /[^\u0000-\u0080]+/.test(path);
+
+  if (isExtendedLengthPath || hasNonAscii) {
+    return path;
+  }
+
+  return path.replace(/\\/g, '/');
+}